+++ js/src/forum/index.ts
@@ -1,3 +1,4 @@
+import CopyButton from "./components/CopyButton";
 import { extend } from "flarum/common/extend";
 import app from "flarum/forum/app";
 import CommentPost from "flarum/forum/components/CommentPost";
@@ -5,45 +6,13 @@ import CommentPost from "flarum/forum/components/CommentPost";
 app.initializers.add("nearata-copy-code-to-clipboard", () => {
   extend(CommentPost.prototype, "oncreate", function () {
     for (const el of this.element.querySelectorAll("pre")) {
-      const copy = document.createElement("button");
+      const container = document.createElement("div");
 
-      copy.classList.add("nearata-copy-code");
-      copy.setAttribute(
-        "title",
-        app.translator.trans("nearata-copy-code-to-clipboard.forum.copy")
-      );
-      copy.style.color = app.forum.attribute("themePrimaryColor");
-      copy.style.borderColor = app.forum.attribute("themePrimaryColor");
+      container.classList.add("NearataCopyCodeToClipboard");
 
-      const copyIcon = document.createElement("i");
-      copyIcon.classList.add("far", "fa-clipboard");
+      el.append(container);
 
-      copy.append(copyIcon);
-      el.prepend(copy);
-    }
-
-    for (const el of this.element.querySelectorAll(".nearata-copy-code")) {
-      el.addEventListener(
-        "click",
-        function (_) {
-          navigator.clipboard.writeText(
-            el.parentNode.querySelector("code").textContent
-          );
-
-          const icon = el.querySelector("i");
-
-          icon.classList.replace("far", "fas");
-          icon.classList.replace("fa-clipboard", "fa-check");
-
-          setTimeout(function () {
-            icon.classList.replace("fas", "far");
-            icon.classList.replace("fa-check", "fa-clipboard");
-
-            clearTimeout(0);
-          }, 1000);
-        },
-        false
-      );
+      m.mount(container, CopyButton);
     }
   });
 });